+++ generators/multipass/voxel_generator_multipass_cb.cpp
@@ -3,6 +3,8 @@
 #include "../../util/containers/container_funcs.h"
 #include "../../util/dstack.h"
 #include "../../util/godot/check_ref_ownership.h"
+#include "../../util/godot/classes/engine.h"
+#include "../../util/godot/classes/script.h"
 #include "../../util/godot/classes/time.h"
 #include "../../util/godot/core/array.h"
 #include "../../util/profiling.h"
@@ -387,6 +389,17 @@ void VoxelGeneratorMultipassCB::clear_cache() {
 	}
 	map.columns.clear();
 	*/
+}
+
+bool VoxelGeneratorMultipassCB::is_runnable() const {
+	Ref<Script> script = get_script();
+	if (script.is_null()) {
+		return false;
+	}
+	if (Engine::get_singleton()->is_editor_hint()) {
+		return script->is_tool();
+	}
+	return true;
 }
 
 bool VoxelGeneratorMultipassCB::debug_try_get_column_states(StdVector<DebugColumnState> &out_states) {