+++ streams/voxel_stream_script.cpp
@@ -2,6 +2,8 @@
 #include "../constants/voxel_string_names.h"
 #include "../storage/voxel_buffer_gd.h"
 #include "../util/godot/check_ref_ownership.h"
+#include "../util/godot/classes/engine.h"
+#include "../util/godot/classes/script.h"
 
 #ifdef ZN_GODOT
 #include "../util/godot/core/class_db.h"
@@ -53,6 +55,17 @@ int VoxelStreamScript::get_used_channels_mask() const {
 		WARN_PRINT_ONCE("VoxelStreamScript::_get_used_channels_mask is unimplemented!");
 	}
 	return mask;
+}
+
+bool VoxelStreamScript::is_runnable() const {
+	Ref<Script> script = get_script();
+	if (script.is_null()) {
+		return false;
+	}
+	if (Engine::get_singleton()->is_editor_hint()) {
+		return script->is_tool();
+	}
+	return true;
 }
 
 void VoxelStreamScript::_bind_methods() {