+++ terrain/variable_lod/voxel_lod_terrain.cpp
@@ -310,20 +310,6 @@ void VoxelLodTerrain::set_stream(Ref<VoxelStream> p_stream) {
 
 	StreamingDependency::reset(_streaming_dependency, p_stream, get_generator());
 
-#ifdef TOOLS_ENABLED
-	if (p_stream.is_valid()) {
-		if (Engine::get_singleton()->is_editor_hint()) {
-			Ref<Script> stream_script = p_stream->get_script();
-			if (stream_script.is_valid()) {
-				// Safety check. It's too easy to break threads by making a script reload.
-				// You can turn it back on, but be careful.
-				_update_data->settings.run_stream_in_editor = false;
-				notify_property_list_changed();
-			}
-		}
-	}
-#endif
-
 	_on_stream_params_changed();
 }
 
@@ -340,20 +326,6 @@ void VoxelLodTerrain::set_generator(Ref<VoxelGenerator> p_generator) {
 
 	MeshingDependency::reset(_meshing_dependency, _mesher, p_generator);
 	StreamingDependency::reset(_streaming_dependency, get_stream(), p_generator);
-
-#ifdef TOOLS_ENABLED
-	if (p_generator.is_valid()) {
-		if (Engine::get_singleton()->is_editor_hint()) {
-			Ref<Script> generator_script = p_generator->get_script();
-			if (generator_script.is_valid()) {
-				// Safety check. It's too easy to break threads by making a script reload.
-				// You can turn it back on, but be careful.
-				_update_data->settings.run_stream_in_editor = false;
-				notify_property_list_changed();
-			}
-		}
-	}
-#endif
 
 	_on_stream_params_changed();
 }
@@ -456,8 +428,7 @@ void VoxelLodTerrain::_on_stream_params_changed() {
 
 	Ref<VoxelGenerator> generator = get_generator();
 
-	if ((stream.is_valid() || generator.is_valid()) &&
-		(Engine::get_singleton()->is_editor_hint() == false || _update_data->settings.run_stream_in_editor)) {
+	if (((stream.is_valid() && stream->is_runnable()) || (generator.is_valid() && generator->is_runnable()))) {
 		start_streamer();
 		start_updater();
 	}
@@ -2670,29 +2641,6 @@ Dictionary VoxelLodTerrain::_b_get_statistics() const {
 	return d;
 }
 
-void VoxelLodTerrain::set_run_stream_in_editor(bool enable) {
-	if (enable == _update_data->settings.run_stream_in_editor) {
-		return;
-	}
-
-	_update_data->wait_for_end_of_task();
-	_update_data->settings.run_stream_in_editor = enable;
-
-	if (Engine::get_singleton()->is_editor_hint()) {
-		if (enable) {
-			_on_stream_params_changed();
-
-		} else {
-			// This is expected to block the main thread until the streaming thread is done.
-			stop_streamer();
-		}
-	}
-}
-
-bool VoxelLodTerrain::is_stream_running_in_editor() const {
-	return _update_data->settings.run_stream_in_editor;
-}
-
 void VoxelLodTerrain::restart_stream() {
 	_on_stream_params_changed();
 }
@@ -3902,9 +3850,6 @@ void VoxelLodTerrain::_bind_methods() {
 	ClassDB::bind_method(D_METHOD("get_voxel_tool"), &Self::get_voxel_tool);
 	ClassDB::bind_method(D_METHOD("save_modified_blocks"), &Self::_b_save_modified_blocks);
 
-	ClassDB::bind_method(D_METHOD("set_run_stream_in_editor"), &Self::set_run_stream_in_editor);
-	ClassDB::bind_method(D_METHOD("is_stream_running_in_editor"), &Self::is_stream_running_in_editor);
-
 	ClassDB::bind_method(D_METHOD("is_area_meshed", "area_in_voxels", "lod_index"), &Self::_b_is_area_meshed);
 
 	// Normalmaps
@@ -4115,12 +4060,6 @@ void VoxelLodTerrain::_bind_methods() {
 
 	ADD_GROUP("Advanced", "");
 
-	// TODO Probably should be in parent class?
-	ADD_PROPERTY(
-			PropertyInfo(Variant::BOOL, "run_stream_in_editor"),
-			"set_run_stream_in_editor",
-			"is_stream_running_in_editor"
-	);
 	ADD_PROPERTY(PropertyInfo(Variant::INT, "mesh_block_size"), "set_mesh_block_size", "get_mesh_block_size");
 	ADD_PROPERTY(
 			PropertyInfo(Variant::BOOL, "full_load_mode_enabled"),